fix react and webpack module resolution and es6
This makes a small change to the entrypoint for rsg to remove an es6
specific syntax (fat arrow).  By making this one entrypoint file
compatible (and using babel-core/register), compatibility can be
extended back to pre-4 versions of node.

This also explicitly makes it so that the webpack build run by RSG look
in this module's node_modules directory for loaders/presets.  Because
the `rsg` command is meant to be run from a parent project, with this as
a 3rd party library, there were issues with webpack searching the parent
- not local - node_modules directory, meaning parent projects would need
to include all of those loaders/presets as direct dependencies.